requirements.txt
discord.py>=2.0
PyNaCl>=1.5
wavelink>=2.0
easyocr>=1.6
aiohttp>=3.0
beautifulsoup4>=4.0
python-dotenv>=0.19
uvloop>=0.19; platform_system != "Windows"
